refactor: save/edit modal + saving functionality
