+++ src/components/molecules/CommandPreview/CommandLabel.styled.tsx
@@ -1,4 +1,4 @@
-import {DeleteOutlined as RawDeleteOutlined} from '@ant-design/icons';
+import {DeleteOutlined as RawDeleteOutlined, EditOutlined as RawEditOutlined} from '@ant-design/icons';
 
 import styled from 'styled-components';
 
@@ -12,8 +12,22 @@ export const LabelContainer = styled.div`
   align-items: center;
 `;
 
+export const ActionsContainer = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 10px;
+`;
+
 export const DeleteOutlined = styled(RawDeleteOutlined)`
   color: ${Colors.red7};
+
+  &:hover {
+    opacity: 0.8;
+  }
+`;
+
+export const EditOutlined = styled(RawEditOutlined)`
+  color: ${Colors.blue6};
 
   &:hover {
     opacity: 0.8;
