+++ src/components/molecules/CommandPreview/SaveCommand.tsx
@@ -1,83 +1,16 @@
 import {useState} from 'react';
 
-import {Button, Input, Modal} from 'antd';
-import {useForm} from 'antd/lib/form/Form';
+import {Button} from 'antd';
 
-import {cloneDeep} from 'lodash';
-import {v4 as uuid} from 'uuid';
-
-import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {updateProjectConfig} from '@redux/reducers/appConfig';
-
-import {Form} from '@components/organisms/TerminalPane/TerminalOptions.styled';
+import SaveEditModal from './SaveEditModal';
 
 const SaveCommand = () => {
-  const dispatch = useAppDispatch();
-
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const [form] = useForm();
 
-  const savedCommandMap = useAppSelector(state => state.config.projectConfig?.savedCommandMap || {});
-
-  const onOkHandler = () => {
-    form.submit();
-  };
-
-  const onSaveCommand = async () => {
-    setIsLoading(true);
-
-    const values: {label: string; content: string} = await form.validateFields();
-    const {label, content} = values;
-
-    const updatedSavedCommandMap = cloneDeep(savedCommandMap);
-    const newCommandId = uuid();
-    updatedSavedCommandMap[newCommandId] = {
-      id: newCommandId,
-      label,
-      content,
-    };
-
-    dispatch(
-      updateProjectConfig({
-        config: {
-          savedCommandMap: updatedSavedCommandMap,
-        },
-        fromConfigFile: false,
-      })
-    );
-    form.resetFields();
-
-    setIsLoading(false);
-    setIsModalOpen(false);
-  };
   return (
     <>
-      <Modal
-        title="Save command"
-        open={isModalOpen}
-        okText="Save"
-        onOk={onOkHandler}
-        onCancel={() => setIsModalOpen(false)}
-        confirmLoading={isLoading}
-        width={800}
-      >
-        <Form form={form} onFinish={onSaveCommand}>
-          <Form.Item
-            name="label"
-            label="Label"
-            rules={[{required: true, message: 'Please provide a label for the command.'}]}
-          >
-            <Input />
-          </Form.Item>
-          <Form.Item name="content" label="Command" rules={[{required: true, message: 'Please provide a command.'}]}>
-            <Input.TextArea
-              autoSize
-              placeholder="Write your shell command here. The command should have YAML output containing k8s resources."
-            />
-          </Form.Item>
-        </Form>
-      </Modal>
+      <SaveEditModal isOpen={isModalOpen} onCancel={() => setIsModalOpen(false)} />
+
       <Button type="link" onClick={() => setIsModalOpen(true)}>
         Save command
       </Button>
